requirements.txt
streamlit
requests
beautifulsoup4
openai
faiss-cpu
numpy
scikit-learn
vaderSentiment
